Use bfloat16 for training and inference
Cast model to bf16 instead of using autocast. Experiment confirmed
negligible accuracy impact (<0.001 val loss diff over 200 steps).
Also add experiment_bf16.py comparing fp32, autocast, and full bf16.

+++ transformer/generate.py
@@ -44,7 +44,7 @@ def main():
     itos = {int(k): v for k, v in json.loads(meta["itos"]).items()}
     vocab_size = int(meta["vocab_size"])
 
-    model = Transformer(vocab_size=vocab_size)
+    model = Transformer(vocab_size=vocab_size).bfloat16()
     model_state = {k[len("model."):]: v for k, v in tensors.items() if k.startswith("model.")}
     model.load_state_dict(model_state, strict=False)
